requirements.txt
pandas
sparkmagic
slackclient
ipython
google-cloud-bigquery
pytest-mock
pytest